Add model package and move classes
