+++ core/src/main/java/bisq/core/app/BisqSetup.java
@@ -25,8 +25,8 @@ import bisq.core.arbitration.ArbitratorManager;
 import bisq.core.arbitration.DisputeManager;
 import bisq.core.btc.listeners.BalanceListener;
 import bisq.core.btc.setup.WalletsSetup;
-import bisq.core.btc.wallet.AddressEntry;
-import bisq.core.btc.wallet.BalanceModel;
+import bisq.core.btc.model.AddressEntry;
+import bisq.core.btc.model.BalanceModel;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.WalletsManager;
 import bisq.core.dao.DaoSetup;
