+++ core/src/main/java/bisq/core/btc/BitcoinModule.java
@@ -22,8 +22,8 @@ import bisq.core.btc.network.BitcoinNodes;
 import bisq.core.btc.network.RegTestHost;
 import bisq.core.btc.network.UserAgent;
 import bisq.core.btc.setup.WalletsSetup;
-import bisq.core.btc.wallet.AddressEntryList;
-import bisq.core.btc.wallet.BalanceModel;
+import bisq.core.btc.model.AddressEntryList;
+import bisq.core.btc.model.BalanceModel;
 import bisq.core.btc.wallet.BsqCoinSelector;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
