+++ core/src/main/java/bisq/core/btc/model/BalanceModel.java
@@ -15,8 +15,9 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.wallet;
+package bisq.core.btc.model;
 
+import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.offer.OpenOfferManager;
 import bisq.core.trade.Trade;
 import bisq.core.trade.TradeManager;
