+++ core/src/main/java/bisq/core/btc/wallet/TradeWalletService.java
@@ -21,6 +21,10 @@ import bisq.core.app.BisqEnvironment;
 import bisq.core.btc.exceptions.SigningException;
 import bisq.core.btc.exceptions.TransactionVerificationException;
 import bisq.core.btc.exceptions.WalletException;
+import bisq.core.btc.model.AddressEntry;
+import bisq.core.btc.model.InputsAndChangeOutput;
+import bisq.core.btc.model.PreparedDepositTxAndMakerInputs;
+import bisq.core.btc.model.RawTransactionInput;
 import bisq.core.btc.setup.WalletConfig;
 import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.locale.Res;
