+++ core/src/main/java/bisq/core/presentation/BalancePresentation.java
@@ -17,7 +17,7 @@
 
 package bisq.core.presentation;
 
-import bisq.core.btc.wallet.BalanceModel;
+import bisq.core.btc.model.BalanceModel;
 import bisq.core.util.BSFormatter;
 
 import javax.inject.Inject;
