+++ core/src/main/java/bisq/core/proto/ProtoDevUtil.java
@@ -18,7 +18,7 @@
 package bisq.core.proto;
 
 import bisq.core.arbitration.DisputeResult;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.offer.AvailabilityResult;
 import bisq.core.offer.Offer;
 import bisq.core.offer.OfferPayload;
