+++ core/src/main/java/bisq/core/trade/TradeManager.java
@@ -18,7 +18,7 @@
 package bisq.core.trade;
 
 import bisq.core.btc.exceptions.AddressEntryException;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
