+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer/BuyerSetupPayoutTxListener.java
@@ -18,7 +18,7 @@
 package bisq.core.trade.protocol.tasks.buyer;
 
 import bisq.core.btc.listeners.AddressConfidenceListener;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.trade.Trade;
 import bisq.core.trade.protocol.tasks.TradeTask;
