+++ core/src/main/java/bisq/core/trade/protocol/tasks/taker/TakerSendPayDepositRequest.java
@@ -17,7 +17,7 @@
 
 package bisq.core.trade.protocol.tasks.taker;
 
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.payment.payload.PaymentAccountPayload;
 import bisq.core.trade.Trade;
