+++ desktop/src/main/java/bisq/desktop/main/funds/deposit/DepositListItem.java
@@ -22,7 +22,7 @@ import bisq.desktop.util.GUIUtil;
 
 import bisq.core.btc.listeners.BalanceListener;
 import bisq.core.btc.listeners.TxConfidenceListener;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.locale.Res;
 import bisq.core.util.BSFormatter;
