+++ desktop/src/main/java/bisq/desktop/main/funds/locked/LockedListItem.java
@@ -20,7 +20,7 @@ package bisq.desktop.main.funds.locked;
 import bisq.desktop.components.AutoTooltipLabel;
 
 import bisq.core.btc.listeners.BalanceListener;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.WalletService;
 import bisq.core.trade.Tradable;
