+++ desktop/src/main/java/bisq/desktop/main/funds/withdrawal/WithdrawalView.java
@@ -30,7 +30,7 @@ import bisq.core.btc.exceptions.AddressEntryException;
 import bisq.core.btc.exceptions.InsufficientFundsException;
 import bisq.core.btc.listeners.BalanceListener;
 import bisq.core.btc.setup.WalletsSetup;
-import bisq.core.btc.wallet.AddressEntry;
+import bisq.core.btc.model.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.Restrictions;
 import bisq.core.locale.Res;
